AI-Negotiation-for-car-lease-or-loan/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]>=2.10,<3
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
httpx==0.27.0
langchain==0.2.0
langchain-groq==0.1.6
google-cloud-vision==3.7.2
pdf2image==1.17.0
pillow==10.3.0
pdfplumber==0.11.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.2
striprtf==0.0.26
beautifulsoup4==4.12.3
extract-msg==0.48.0
pillow-heif==0.16.0
pytesseract==0.3.10
psycopg2-binary==2.9.9
python-dotenv==1.0.1
qdrant-client==1.9.1
fastembed==0.4.1
langchain-community==0.2.0
